Run 2 test jobs in parallel

+++ .travis/run_integration_test_2.sh
@@ -1,19 +1,19 @@
 #!/usr/bin/env bash
 set -e
 
-#if [ ! -f $HOME/.m2/repository/org/rocksdb/rocksdbjni/6.19.3/rocksdbjni-6.19.3.jar ]; then
+if [ ! -f $HOME/.m2/repository/org/rocksdb/rocksdbjni/6.19.3/rocksdbjni-6.19.3.jar ]; then
     echo "Put locally built rocksdbjni-6.19.3.jar in the local maven repository"
     mkdir -p $HOME/.m2/repository/org/rocksdb/rocksdbjni/6.19.3
     mv -f rocksdbjni-6.19.3.jar $HOME/.m2/repository/org/rocksdb/rocksdbjni/6.19.3/rocksdbjni-6.19.3.jar
     mv -f rocksdbjni-6.19.3.jar.sha1 $HOME/.m2/repository/org/rocksdb/rocksdbjni/6.19.3/rocksdbjni-6.19.3.jar.sha1
-#fi
+fi
 
-#if [ ! -f $HOME/.m2/repository/org/rocksdb/rocksdbjni/6.22.1.1/rocksdbjni-6.22.1.1.jar ]; then
+if [ ! -f $HOME/.m2/repository/org/rocksdb/rocksdbjni/6.22.1.1/rocksdbjni-6.22.1.1.jar ]; then
     echo "Put locally built rocksdbjni-6.22.1.1.jar in the local maven repository"
     mkdir -p $HOME/.m2/repository/org/rocksdb/rocksdbjni/6.22.1.1
     mv -f rocksdbjni-6.22.1.1.jar $HOME/.m2/repository/org/rocksdb/rocksdbjni/6.22.1.1/rocksdbjni-6.22.1.1.jar
     mv -f rocksdbjni-6.22.1.1.jar.sha1 $HOME/.m2/repository/org/rocksdb/rocksdbjni/6.22.1.1/rocksdbjni-6.22.1.1.jar.sha1
-#fi
+fi
 
 export PATH=$HOME/apache-maven-3.8.2/bin:$PATH
